Refactor package manager to avoid circular deps
