+++ components/os/generic_packagemanager.go
@@ -1,4 +1,4 @@
-package command
+package os
 
 import (
 	"fmt"
@@ -6,14 +6,14 @@ import (
 	"github.com/DataDog/test-infra-definitions/common"
 	"github.com/DataDog/test-infra-definitions/common/namer"
 	"github.com/DataDog/test-infra-definitions/common/utils"
-	"github.com/DataDog/test-infra-definitions/components/os"
+	"github.com/DataDog/test-infra-definitions/components/command"
 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 )
 
 type GenericPackageManager struct {
 	namer           namer.Namer
-	updateDBCommand Command
-	runner          Runner
+	updateDBCommand command.Command
+	runner          command.Runner
 	opts            []pulumi.ResourceOption
 	installCmd      string
 	updateCmd       string
@@ -21,7 +21,7 @@ type GenericPackageManager struct {
 }
 
 func NewGenericPackageManager(
-	runner Runner,
+	runner command.Runner,
 	name string,
 	installCmd string,
 	updateCmd string,
@@ -38,8 +38,8 @@ func NewGenericPackageManager(
 	return packageManager
 }
 
-func (m *GenericPackageManager) Ensure(packageRef string, transform Transformer, checkBinary string, opts ...os.PackageManagerOption) (Command, error) {
-	params, err := common.ApplyOption(&os.PackageManagerParams{}, opts)
+func (m *GenericPackageManager) Ensure(packageRef string, transform command.Transformer, checkBinary string, opts ...PackageManagerOption) (command.Command, error) {
+	params, err := common.ApplyOption(&PackageManagerParams{}, opts)
 	if err != nil {
 		return nil, err
 	}
@@ -61,7 +61,7 @@ func (m *GenericPackageManager) Ensure(packageRef string, transform Transformer,
 	}
 
 	cmdName := m.namer.ResourceName("install-"+packageRef, utils.StrHash(cmdStr))
-	var cmdArgs RunnerCommandArgs = &Args{
+	var cmdArgs command.RunnerCommandArgs = &command.Args{
 		Create:      pulumi.String(cmdStr),
 		Environment: m.env,
 		Sudo:        true,
@@ -82,14 +82,14 @@ func (m *GenericPackageManager) Ensure(packageRef string, transform Transformer,
 	return cmd, nil
 }
 
-func (m *GenericPackageManager) updateDB(opts []pulumi.ResourceOption) (Command, error) {
+func (m *GenericPackageManager) updateDB(opts []pulumi.ResourceOption) (command.Command, error) {
 	if m.updateDBCommand != nil {
 		return m.updateDBCommand, nil
 	}
 
 	c, err := m.runner.Command(
 		m.namer.ResourceName("update"),
-		&Args{
+		&command.Args{
 			Create:      pulumi.String(m.updateCmd),
 			Environment: m.env,
 			Sudo:        true,
