+++ components/os/linux_packagemanagers.go
@@ -6,14 +6,14 @@ import (
 )
 
 func newAptManager(runner command.Runner) PackageManager {
-	return command.NewGenericPackageManager(runner, "apt", "apt-get install -y", "apt-get update -y",
+	return NewGenericPackageManager(runner, "apt", "apt-get install -y", "apt-get update -y",
 		pulumi.StringMap{"DEBIAN_FRONTEND": pulumi.String("noninteractive")})
 }
 
 func newYumManager(runner command.Runner) PackageManager {
-	return command.NewGenericPackageManager(runner, "yum", "yum install -y", "", nil)
+	return NewGenericPackageManager(runner, "yum", "yum install -y", "", nil)
 }
 
 func newDnfManager(runner command.Runner) PackageManager {
-	return command.NewGenericPackageManager(runner, "dnf", "dnf install -y", "", nil)
+	return NewGenericPackageManager(runner, "dnf", "dnf install -y", "", nil)
 }
