+++ components/os/macos_packagemanagers.go
@@ -6,6 +6,6 @@ import (
 )
 
 func newBrewManager(runner command.Runner) PackageManager {
-	return command.NewGenericPackageManager(runner, "brew", "brew install -y", "brew update -y",
+	return NewGenericPackageManager(runner, "brew", "brew install -y", "brew update -y",
 		pulumi.StringMap{"NONINTERACTIVE": pulumi.String("1")})
 }
